Refactor: python: Split base test classes into their own files.
No significant code changes here - it's mostly just changing lists of
imports and pylint suppressions to suit what each new file needs.
