+++ python/pacemaker/_cts/tests/__init__.py
@@ -4,3 +4,10 @@ Test classes for the `pacemaker._cts` package.
 
 __copyright__ = "Copyright 2023 the Pacemaker project contributors"
 __license__ = "GNU Lesser General Public License version 2.1 or later (LGPLv2.1+)"
+
+from pacemaker._cts.tests.ctstest import CTSTest
+from pacemaker._cts.tests.remotedriver import RemoteDriver
+from pacemaker._cts.tests.simulstartlite import SimulStartLite
+from pacemaker._cts.tests.simulstoplite import SimulStopLite
+from pacemaker._cts.tests.starttest import StartTest
+from pacemaker._cts.tests.stoptest import StopTest
