+++ python/pacemaker/_cts/tests/remotedriver.py
@@ -1,22 +1,19 @@
 """ Base classes for CTS tests """
 
-__all__ = ["CTSTest", "RemoteDriver", "SimulStartLite", "SimulStopLite", "StartTest", "StopTest"]
+__all__ = ["RemoteDriver"]
 __copyright__ = "Copyright 2000-2023 the Pacemaker project contributors"
 __license__ = "GNU General Public License version 2 or later (GPLv2+) WITHOUT ANY WARRANTY"
 
 import os
-import re
 import time
 import subprocess
 import tempfile
 
-from pacemaker._cts.audits import AuditConstraint, AuditResource
-from pacemaker._cts.environment import EnvFactory
-from pacemaker._cts.logging import LogFactory
-from pacemaker._cts.patterns import PatternSelector
-from pacemaker._cts.remote import RemoteFactory
+from pacemaker._cts.tests.ctstest import CTSTest
+from pacemaker._cts.tests.simulstartlite import SimulStartLite
+from pacemaker._cts.tests.starttest import StartTest
+from pacemaker._cts.tests.stoptest import StopTest
 from pacemaker._cts.timer import Timer
-from pacemaker._cts.watcher import LogWatcher
 
 # Disable various pylint warnings that occur in so many places throughout this
 # file it's easiest to just take care of them globally.  This does introduce the
@@ -25,275 +22,6 @@ from pacemaker._cts.watcher import LogWatcher
 
 # pylint doesn't understand that self._rsh is callable.
 # pylint: disable=not-callable
-# pylint doesn't understand that self._env is subscriptable.
-# pylint: disable=unsubscriptable-object
-
-
-class CTSTest:
-    """ The base class for all cluster tests.  This implements a basic set of
-        properties and behaviors like setup, tear down, time keeping, and
-        statistics tracking.  It is up to specific tests to implement their own
-        specialized behavior on top of this class.
-    """
-
-    def __init__(self, cm):
-        """ Create a new CTSTest instance
-
-            Arguments:
-
-            cm -- A ClusterManager instance
-        """
-
-        # pylint: disable=invalid-name
-
-        self.audits = []
-        self.name = None
-        self.templates = PatternSelector(cm["Name"])
-
-        self.stats = { "auditfail": 0,
-                      "calls": 0,
-                      "failure": 0,
-                      "skipped": 0,
-                      "success": 0 }
-
-        self._cm = cm
-        self._env = EnvFactory().getInstance()
-        self._r_o2cb = None
-        self._r_ocfs2 = []
-        self._rsh = RemoteFactory().getInstance()
-        self._logger = LogFactory()
-        self._timers = {}
-
-        self.benchmark = True  # which tests to benchmark
-        self.failed = False
-        self.is_container = False
-        self.is_experimental = False
-        self.is_loop = False
-        self.is_unsafe = False
-        self.is_valgrind = False
-        self.passed = True
-
-    def log(self, args):
-        """ Log a message """
-
-        self._logger.log(args)
-
-    def debug(self, args):
-        """ Log a debug message """
-
-        self._logger.debug(args)
-
-    def get_timer(self, key="test"):
-        """ Get the start time of the given timer """
-
-        try:
-            return self._timers[key].start_time
-        except KeyError:
-            return 0
-
-    def set_timer(self, key="test"):
-        """ Set the start time of the given timer to now, and return
-            that time
-        """
-
-        if key not in self._timers:
-            self._timers[key] = Timer(self._logger, self.name, key)
-
-        self._timers[key].start()
-        return self._timers[key].start_time
-
-    def log_timer(self, key="test"):
-        """ Log the elapsed time of the given timer """
-
-        if key not in self._timers:
-            return
-
-        elapsed = self._timers[key].elapsed
-        self.debug("%s:%s runtime: %.2f" % (self.name, key, elapsed))
-        del self._timers[key]
-
-    def incr(self, name):
-        """ Increment the given stats key """
-
-        if name not in self.stats:
-            self.stats[name] = 0
-
-        self.stats[name] += 1
-
-        # Reset the test passed boolean
-        if name == "calls":
-            self.passed = True
-
-    def failure(self, reason="none"):
-        """ Increment the failure count, with an optional failure reason """
-
-        self.passed = False
-        self.incr("failure")
-        self._logger.log(("Test %s" % self.name).ljust(35) + " FAILED: %s" % reason)
-
-        return False
-
-    def success(self):
-        """ Increment the success count """
-
-        self.incr("success")
-        return True
-
-    def skipped(self):
-        """ Increment the skipped count """
-
-        self.incr("skipped")
-        return True
-
-    def __call__(self, node):
-        """ Perform this test """
-
-        raise NotImplementedError
-
-    def audit(self):
-        """ Perform all the relevant audits (see ClusterAudit), returning
-            whether or not they all passed.
-        """
-
-        passed = True
-
-        for audit in self.audits:
-            if not audit():
-                self._logger.log("Internal %s Audit %s FAILED." % (self.name, audit.name))
-                self.incr("auditfail")
-                passed = False
-
-        return passed
-
-    def setup(self, node):
-        """ Setup this test """
-
-        # node is used in subclasses
-        # pylint: disable=unused-argument
-
-        return self.success()
-
-    def teardown(self, node):
-        """ Tear down this test """
-
-        # node is used in subclasses
-        # pylint: disable=unused-argument
-
-        return self.success()
-
-    def create_watch(self, patterns, timeout, name=None):
-        """ Create a new LogWatcher object with the given patterns, timeout,
-            and optional name.  This object can be used to search log files
-            for matching patterns during this test's run.
-        """
-        if not name:
-            name = self.name
-
-        return LogWatcher(self._env["LogFileName"], patterns, self._env["nodes"], self._env["LogWatcher"], name, timeout)
-
-    def local_badnews(self, prefix, watch, local_ignore=None):
-        """ Use the given watch object to search through log files for messages
-            starting with the given prefix.  If no prefix is given, use
-            "LocalBadNews:" by default.  The optional local_ignore list should
-            be a list of regexes that, if found in a line, will cause that line
-            to be ignored.
-
-            Return the number of matches found.
-        """
-        errcount = 0
-        if not prefix:
-            prefix = "LocalBadNews:"
-
-        ignorelist = [" CTS: ", prefix]
-
-        if local_ignore:
-            ignorelist += local_ignore
-
-        while errcount < 100:
-            match = watch.look(0)
-            if match:
-                add_err = True
-
-                for ignore in ignorelist:
-                    if add_err and re.search(ignore, match):
-                        add_err = False
-
-                if add_err:
-                    self._logger.log("%s %s" % (prefix, match))
-                    errcount += 1
-            else:
-                break
-        else:
-            self._logger.log("Too many errors!")
-
-        watch.end()
-        return errcount
-
-    def is_applicable(self):
-        """ Return True if this test is applicable in the current test configuration.
-            This method must be implemented by all subclasses.
-        """
-
-        if self.is_loop and not self._env["loop-tests"]:
-            return False
-
-        if self.is_unsafe and not self._env["unsafe-tests"]:
-            return False
-
-        if self.is_valgrind and not self._env["valgrind-tests"]:
-            return False
-
-        if self.is_experimental and not self._env["experimental-tests"]:
-            return False
-
-        if self.is_container and not self._env["container-tests"]:
-            return False
-
-        if self._env["benchmark"] and not self.benchmark:
-            return False
-
-        return True
-
-    def _find_ocfs2_resources(self, node):
-        """ Find any OCFS2 filesystems mounted on the given cluster node,
-            populating the internal self._r_ocfs2 list with them and returning
-            the number of OCFS2 filesystems.
-         """
-
-        self._r_o2cb = None
-        self._r_ocfs2 = []
-
-        (_, lines) = self._rsh(node, "crm_resource -c", verbose=1)
-        for line in lines:
-            if re.search("^Resource", line):
-                r = AuditResource(self._cm, line)
-
-                if r.rtype == "o2cb" and r.parent != "NA":
-                    self.debug("Found o2cb: %s" % self._r_o2cb)
-                    self._r_o2cb = r.parent
-
-            if re.search("^Constraint", line):
-                c = AuditConstraint(self._cm, line)
-
-                if c.type == "rsc_colocation" and c.target == self._r_o2cb:
-                    self._r_ocfs2.append(c.rsc)
-
-        self.debug("Found ocfs2 filesystems: %s" % self._r_ocfs2)
-        return len(self._r_ocfs2)
-
-    def can_run_now(self, node):
-        """ Return True if we can meaningfully run right now """
-
-        # node is used in subclasses
-        # pylint: disable=unused-argument
-
-        return True
-
-    @property
-    def errors_to_ignore(self):
-        """ Return list of errors which should be ignored """
-
-        return []
 
 
 class RemoteDriver(CTSTest):
@@ -803,312 +531,3 @@ class RemoteDriver(CTSTest):
         return [ r"""is running on remote.*which isn't allowed""",
                  r"""Connection terminated""",
                  r"""Could not send remote""" ]
-
-
-class SimulStartLite(CTSTest):
-    """ A pseudo-test that is only used to set up conditions before running
-        some other test.  This class starts any stopped nodes more or less
-        simultaneously.
-
-        Other test classes should not use this one as a superclass.
-    """
-
-    def __init__(self, cm):
-        """ Create a new SimulStartLite instance
-
-            Arguments:
-
-            cm -- A ClusterManager instance
-        """
-
-        CTSTest.__init__(self,cm)
-        self.name = "SimulStartLite"
-
-    def __call__(self, dummy):
-        """ Start all stopped nodes more or less simultaneously, returning
-            whether this succeeded or not.
-        """
-
-        self.incr("calls")
-        self.debug("Setup: %s" % self.name)
-
-        # We ignore the "node" parameter...
-        node_list = []
-        for node in self._env["nodes"]:
-            if self._cm.ShouldBeStatus[node] == "down":
-                self.incr("WasStopped")
-                node_list.append(node)
-
-        self.set_timer()
-        while len(node_list) > 0:
-            # Repeat until all nodes come up
-            uppat = self.templates["Pat:NonDC_started"]
-            if self._cm.upcount() == 0:
-                uppat = self.templates["Pat:Local_started"]
-
-            watchpats = [ self.templates["Pat:DC_IDLE"] ]
-            for node in node_list:
-                watchpats.extend([uppat % node,
-                                  self.templates["Pat:InfraUp"] % node,
-                                  self.templates["Pat:PacemakerUp"] % node])
-
-            #   Start all the nodes - at about the same time...
-            watch = self.create_watch(watchpats, self._env["DeadTime"]+10)
-            watch.set_watch()
-
-            stonith = self._cm.prepare_fencing_watcher(self.name)
-
-            for node in node_list:
-                self._cm.StartaCMnoBlock(node)
-
-            watch.look_for_all()
-
-            node_list = self._cm.fencing_cleanup(self.name, stonith)
-
-            if node_list is None:
-                return self.failure("Cluster did not stabilize")
-
-            # Remove node_list messages from watch.unmatched
-            for node in node_list:
-                self._logger.debug("Dealing with stonith operations for %s" % node_list)
-                if watch.unmatched:
-                    try:
-                        watch.unmatched.remove(uppat % node)
-                    except ValueError:
-                        self.debug("Already matched: %s" % (uppat % node))
-
-                    try:
-                        watch.unmatched.remove(self.templates["Pat:InfraUp"] % node)
-                    except ValueError:
-                        self.debug("Already matched: %s" % (self.templates["Pat:InfraUp"] % node))
-
-                    try:
-                        watch.unmatched.remove(self.templates["Pat:PacemakerUp"] % node)
-                    except ValueError:
-                        self.debug("Already matched: %s" % (self.templates["Pat:PacemakerUp"] % node))
-
-            if watch.unmatched:
-                for regex in watch.unmatched:
-                    self._logger.log ("Warn: Startup pattern not found: %s" % regex)
-
-            if not self._cm.cluster_stable():
-                return self.failure("Cluster did not stabilize")
-
-        did_fail = False
-        unstable = []
-        for node in self._env["nodes"]:
-            if self._cm.StataCM(node) == 0:
-                did_fail = True
-                unstable.append(node)
-
-        if did_fail:
-            return self.failure("Unstarted nodes exist: %s" % unstable)
-
-        unstable = []
-        for node in self._env["nodes"]:
-            if not self._cm.node_stable(node):
-                did_fail = True
-                unstable.append(node)
-
-        if did_fail:
-            return self.failure("Unstable cluster nodes exist: %s" % unstable)
-
-        return self.success()
-
-    def is_applicable(self):
-        """ SimulStartLite is a setup test and never applicable """
-
-        return False
-
-
-class SimulStopLite(CTSTest):
-    """ A pseudo-test that is only used to set up conditions before running
-        some other test.  This class stops any running nodes more or less
-        simultaneously.  It can be used both to set up a test or to clean up
-        a test.
-
-        Other test classes should not use this one as a superclass.
-    """
-
-    def __init__(self, cm):
-        """ Create a new SimulStopLite instance
-
-            Arguments:
-
-            cm -- A ClusterManager instance
-        """
-
-        CTSTest.__init__(self,cm)
-        self.name = "SimulStopLite"
-
-    def __call__(self, dummy):
-        """ Stop all running nodes more or less simultaneously, returning
-            whether this succeeded or not.
-        """
-
-        self.incr("calls")
-        self.debug("Setup: %s" % self.name)
-
-        # We ignore the "node" parameter...
-        watchpats = []
-
-        for node in self._env["nodes"]:
-            if self._cm.ShouldBeStatus[node] == "up":
-                self.incr("WasStarted")
-                watchpats.append(self.templates["Pat:We_stopped"] % node)
-
-        if len(watchpats) == 0:
-            return self.success()
-
-        # Stop all the nodes - at about the same time...
-        watch = self.create_watch(watchpats, self._env["DeadTime"]+10)
-
-        watch.set_watch()
-        self.set_timer()
-        for node in self._env["nodes"]:
-            if self._cm.ShouldBeStatus[node] == "up":
-                self._cm.StopaCMnoBlock(node)
-
-        if watch.look_for_all():
-            # Make sure they're completely down with no residule
-            for node in self._env["nodes"]:
-                self._rsh(node, self.templates["StopCmd"])
-
-            return self.success()
-
-        did_fail = False
-        up_nodes = []
-        for node in self._env["nodes"]:
-            if self._cm.StataCM(node) == 1:
-                did_fail = True
-                up_nodes.append(node)
-
-        if did_fail:
-            return self.failure("Active nodes exist: %s" % up_nodes)
-
-        self._logger.log("Warn: All nodes stopped but CTS didn't detect: %s" % watch.unmatched)
-        return self.failure("Missing log message: %s " % watch.unmatched)
-
-    def is_applicable(self):
-        """ SimulStopLite is a setup test and never applicable """
-
-        return False
-
-
-class StartTest(CTSTest):
-    """ A pseudo-test that is only used to set up conditions before running
-        some other test.  This class starts the cluster manager on a given
-        node.
-
-        Other test classes should not use this one as a superclass.
-    """
-
-    def __init__(self, cm):
-        """ Create a new StartTest instance
-
-            Arguments:
-
-            cm -- A ClusterManager instance
-        """
-
-        CTSTest.__init__(self,cm)
-        self.name = "Start"
-
-    def __call__(self, node):
-        """ Start the given node, returning whether this succeeded or not """
-
-        self.incr("calls")
-
-        if self._cm.upcount() == 0:
-            self.incr("us")
-        else:
-            self.incr("them")
-
-        if self._cm.ShouldBeStatus[node] != "down":
-            return self.skipped()
-
-        if self._cm.StartaCM(node):
-            return self.success()
-
-        return self.failure("Startup %s on node %s failed"
-                            % (self._env["Name"], node))
-
-
-class StopTest(CTSTest):
-    """ A pseudo-test that is only used to set up conditions before running
-        some other test.  This class stops the cluster manager on a given
-        node.
-
-        Other test classes should not use this one as a superclass.
-    """
-
-    def __init__(self, cm):
-        """ Create a new StopTest instance
-
-            Arguments:
-
-            cm -- A ClusterManager instance
-        """
-
-        CTSTest.__init__(self, cm)
-        self.name = "Stop"
-
-    def __call__(self, node):
-        """ Stop the given node, returning whether this succeeded or not """
-
-        self.incr("calls")
-        if self._cm.ShouldBeStatus[node] != "up":
-            return self.skipped()
-
-        # Technically we should always be able to notice ourselves stopping
-        patterns = [ self.templates["Pat:We_stopped"] % node ]
-
-        # Any active node needs to notice this one left
-        # (note that this won't work if we have multiple partitions)
-        for other in self._env["nodes"]:
-            if self._cm.ShouldBeStatus[other] == "up" and other != node:
-                patterns.append(self.templates["Pat:They_stopped"] %(other, self._cm.key_for_node(node)))
-
-        watch = self.create_watch(patterns, self._env["DeadTime"])
-        watch.set_watch()
-
-        if node == self._cm.OurNode:
-            self.incr("us")
-        else:
-            if self._cm.upcount() <= 1:
-                self.incr("all")
-            else:
-                self.incr("them")
-
-        self._cm.StopaCM(node)
-        watch.look_for_all()
-
-        failreason = None
-        unmatched_str = "||"
-
-        if watch.unmatched:
-            (_, output) = self._rsh(node, "/bin/ps axf", verbose=1)
-            for line in output:
-                self.debug(line)
-
-            (_, output) = self._rsh(node, "/usr/sbin/dlm_tool dump 2>/dev/null", verbose=1)
-            for line in output:
-                self.debug(line)
-
-            for regex in watch.unmatched:
-                self._logger.log ("ERROR: Shutdown pattern not found: %s" % regex)
-                unmatched_str +=  "%s||" % regex
-                failreason = "Missing shutdown pattern"
-
-        self._cm.cluster_stable(self._env["DeadTime"])
-
-        if not watch.unmatched or self._cm.upcount() == 0:
-            return self.success()
-
-        if len(watch.unmatched) >= self._cm.upcount():
-            return self.failure("no match against (%s)" % unmatched_str)
-
-        if failreason is None:
-            return self.success()
-
-        return self.failure(failreason)
